% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapArrange.R
\name{heatmapArrange}
\alias{heatmapArrange}
\title{heatmapArrange}
\usage{
heatmapArrange(data_in, class_by = "", cluster_row = FALSE,
  cluster_column = FALSE, scale = TRUE, by_row = TRUE)
}
\arguments{
\item{data_in}{is a dataframe of gene expression data}

\item{class_by}{is a vector of strings/regular expression that are used to
parse column names on the fly to produce a class column in the dataframe 
output. These must reflect some common element within the input data's sample
names that can be used to define classes. See examples...}

\item{cluster_row}{A logical value for whether row clustering is performed}

\item{cluster_column}{A logical value for whether column clustering is 
performed}

\item{scale}{A logical value for whether scaling is performed}

\item{by_row}{A logical value for whether scaling is performed by row (TRUE)
of column (FALSE)}

\item{by_row}{determines whether scaling is performed on rows or columns}
}
\value{
A dataframe compatible with ggplot2
}
\description{
A function to take a matrix/dataframe of gene expression data and return a 
dataframe suitable for plotting as a heatmap with ggplot2
}
\examples{
#synthesise example data matrix
mtx <- matrix(rnorm(mean = 7, n = 90), ncol = 10) 
row.names(mtx) <- LETTERS[1:9]
colnames(mtx) <- c(paste0("X", 1:5), paste0("Y", 6:10))

hm_dfr <- heatmapArrange(mtx, 
                         class_by = c("X", "Y"),
                         cluster_row = TRUE,
                         cluster_column = TRUE,
                         scale = TRUE,
                         by_row = TRUE
) 

#library(ggplot2)
#ggplot(hm_dfr, aes(x = x, y  = y, fill = )) 

}
