% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idConvert.R
\name{idConvert}
\alias{idConvert}
\title{idConvert}
\usage{
idConvert(ids, format_in = "ensembl_gene_id", format_out = "hgnc_symbol")
}
\arguments{
\item{ids}{A vector of genomic identifiers}

\item{format_in}{Format/type of genomic identifiers of the ids input}

\item{format_out}{Format/type of genomic identifiers to be returned}
}
\value{
A 2-column dataframe of original and converted ids
}
\description{
Provided with a character vector of genomic identifiers; and a biomaRt
hsapiens_gene_ensembl specific set of filters and attributes (i.e. the
id format of the input and output respectively); this function returns
a data frame with the corresponding mappings of the given identifiers
to each of specified attributes.
}
\details{
The 'format_in' argument defaults to ensembl gene id, and the mapped
'format_out' argument defaults to hgnc symbol.

Common filters/attributes are "ensembl_gene_id", "hgnc_symbol", 
"entrezgene", "illumina_humanht_12_v4" and "affy_hg_u133a".

Credit to Gil Tomas for the original code. Requires biomaRt and 
internet access.
}
\examples{
id_vec <- c("ESR1", "ERBB2", "AURKA")

idConvert(ids = id_vec,
                format_in = "hgnc_symbol",
                format_out = "entrezgene"
                )

}
