% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsArrange.R
\name{mdsArrange}
\alias{mdsArrange}
\title{mdsArrange}
\usage{
mdsArrange(d, isAlreadyScaled = FALSE)
}
\arguments{
\item{d}{is a dataframe of gene expression data}

\item{isAlreadyScaled}{Logical value for whether d is already scaled}
}
\value{
A dataframe compatible with ggplot2
}
\description{
A function to take a matrix/dataframe of gene expression data and return a 
dataframe suitable for plotting as an MDS plot with ggplot2
}
\examples{
#synthesise example data matrix
mtx <- matrix(rnorm(mean = 7, n = 90), ncol = 10) 
row.names(mtx) <- LETTERS[1:9]
colnames(mtx) <- c(paste0("X", 1:5), paste0("Y", 6:10))

hm_dfr <- mdsArrange(d = mtx, isAlreadyScaled = FALSE)

#library(ggplot2)
#ggplot(hm_dfr, aes(x = x, y  = y, fill = )) 

}
