% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostVar.R
\name{mostVar}
\alias{mostVar}
\title{mostVar
Calculate the top n most variable genes in a matrix of gene expression data}
\usage{
mostVar(data, n, i_want_most_var = TRUE)
}
\arguments{
\item{data}{An expression matrix (samples in columns, genes in rows)}

\item{n}{The number of variable genes to return}

\item{i_want_most_var}{Logical value to return most variable (TRUE) or least
variable (FALSE) genes}
}
\value{
A subset of n most variable rows from the original data
}
\description{
mostVar
Calculate the top n most variable genes in a matrix of gene expression data
}
\examples{
#synthesise example data matrix
mtx <- matrix(rnorm(mean = 7, n = 90), ncol = 10) 
row.names(mtx) <- LETTERS[1:9]
colnames(mtx) <- c(paste0("X", 1:5), paste0("Y", 6:10))

mv_3 <- mostVar(data = mtx, n = 3)

}
