% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nki.R
\docType{data}
\name{nki}
\alias{nki}
\title{NKI invasive breast cancer expressionSet}
\format{An expressionSet with 337 samples and 24481 features, as well as 
phenotypic and feature data}
\source{
\url{http://bioconductor.org/packages/release/data/experiment/html/breastCancerNKI.html}
}
\usage{
nki
}
\description{
Originally used as a[validation dataset for the 70-gene MammaPrint signature
(https://www.ncbi.nlm.nih.gov/pubmed/12490681) this dataset has grown 
slightly in size but still is effectively the same collection of invasive 
breast cancer samples of no specific subtype. More information can be found 
in the bioconductor package's web page but, because there are issues with 
installing this on windows I've simply included the data here as part of 
transcripTools.
}
\keyword{datasets}
