% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickBuildeSet.R
\name{quickBuildeSet}
\alias{quickBuildeSet}
\title{quickBuildeSet}
\usage{
quickBuildeSet(xpr, pheno)
}
\arguments{
\item{xpr}{Expression matrix, samples in columns, genes in rows. Column order
must be identical to pheno row order}

\item{pheno}{Phenotypic data matrix, samples in rows, variables in columns. 
Row order must be identical to xpr column order}
}
\value{
An eSet of the combine matrices
}
\description{
Build an minimal expressionSet from matched expression and phenotypic 
matrices. Caution should be employed as this eSet will lack feature data and
proper phenotypic data variable descriptions.
}
\examples{
# Here we'll first dismantle the NKI dataset before re-combining as a minimal
# eSet
library(breastCancerNKI)
library(Biobase)
data(nki)
xpr <- exprs(nki)
pheno <- pData(nki)

minimal_nki <- quickBuildeSet(xpr, pheno)
# note that feature data and variable descriptions are now missing
}
