% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses_GO.R
\name{up_go}
\alias{up_go}
\title{Perform GO Analysis on Up-regulated Genes}
\usage{
up_go(
  df,
  samples,
  databases = c("MSigDB_Hallmark_2020", "GO_Biological_Process_2023", "BioPlanet_2019",
    "GO_Cellular_Component_2023", "Reactome_Pathways_2024")
)
}
\arguments{
\item{df}{Data frame of DE results with rownames = symbols and a \code{logFC} column.}

\item{samples}{Character used in plot titles.}

\item{databases}{Character vector of Enrichr databases to use (any length).}
}
\value{
Named list of ggplot objects (one per database).
}
\description{
This function performs Gene Ontology (GO) analysis on up-regulated genes from a differential
expression data frame. By default, it uses a preset list of GO and pathway databases. However,
you can supply your own database vector via the \code{databases} parameter. The function also
creates bar plots for the top 20 GO terms (based on p-value) for each database.
}
