% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsArrange.R
\name{mdsArrange}
\alias{mdsArrange}
\title{mdsArrange}
\usage{
mdsArrange(d, isAlreadyScaled = FALSE)
}
\arguments{
\item{d}{is a dataframe of gene expression data}

\item{isAlreadyScaled}{Logical value for whether d is already scaled}
}
\value{
A dataframe compatible with ggplot2
}
\description{
A function to take a matrix/dataframe of gene expression data and return a 
dataframe suitable for plotting as an MDS plot with ggplot2
}
\examples{
#Using the NKI dataset
library(breastCancerNKI)
library(Biobase)
data(nki)
mtx <- exprs(nki)

#Calculate PCs 1 & 2
mds_dfr <- mdsArrange(d = mtx, isAlreadyScaled = TRUE)
#Merge with phenotypic data for plotting
mds_mrg <- merge(mds_dfr, pData(nki), by = 0)

#Plot MDS and colour by estrogen receptor positivity
library(ggplot2)
ggplot(mds_mrg, aes(x = x, y  = y, colour = as.factor(er))) + 
    geom_point()

}
