% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-build-expression-set.R
\name{quickbuildeSet}
\alias{quickbuildeSet}
\title{quickBuildeSet
Build an expressionSet from matched expression and phenotypic matrices}
\usage{
quickbuildeSet(xpr, pheno)
}
\arguments{
\item{xpr}{Expression matrix, samples in columns, genes in rows. Column order
must be identical to pheno row order}

\item{pheno}{Phenotypic data matrix, samples in rows, variables in columns. 
Row order must be identical to xpr column order}
}
\value{
An eSet of the combine matrices
}
\description{
quickBuildeSet
Build an expressionSet from matched expression and phenotypic matrices
}
